Move AuthorDto
