+++ src/main/java/net/kosa/mentopingserver/domain/post/dto/QuestionResponseDto.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import net.kosa.mentopingserver.domain.mentor.dto.AuthorDto;
+import net.kosa.mentopingserver.domain.member.dto.AuthorDto;
 
 import java.time.LocalDateTime;
 import java.util.List;
